'wrote a new dataset builder to take stored files from teh physionet 2021 challenge and convert them into a sections of 12x1beat I also imported the helper_code.py from physionet to aid my work'

+++ src/ptb/ptb_dataset_builder.py
@@ -30,7 +30,7 @@ class Builder(tfds.core.GeneratorBasedBuilder):
                 }),
                 'subject': np.int32,
                 'quality': tfds.features.ClassLabel(names=['Unacceptable', 'Barely acceptable', 'Excellent', '[]']),
-                'age': np.uint8,
+                'age': np.uint16,
                 'gender': np.uint8,
                 'diagnostic': tfds.features.ClassLabel(names=['STTC', 'NORM', 'MI', 'HYP', 'CD', 'NAV']),
             }),
